refactor: extract Tab practice into one folder
